# cli/src/pcluster/api/openapi/openapi.yaml
openapi: 3.0.2
info:
  description: ParallelCluster API
  title: ParallelCluster
  version: 3.11.0
servers:
- url: /
security:
- aws.auth.sigv4: []
paths:
  /v3/clusters:
    get:
      description: Retrieve the list of existing clusters.
      operationId: list_clusters
      parameters:
      - description: List clusters deployed to a given AWS Region.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: List clusters deployed to a given AWS Region.
          type: string
        style: form
      - description: Token to use for paginated requests.
        explode: true
        in: query
        name: nextToken
        required: false
        schema:
          description: Token to use for paginated requests.
          type: string
        style: form
      - description: Filter by cluster status. (Defaults to all clusters.)
        explode: true
        in: query
        name: clusterStatus
        required: false
        schema:
          description: Filter by cluster status. (Defaults to all clusters.)
          items:
            $ref: '#/components/schemas/ClusterStatusFilteringOption'
          type: array
          uniqueItems: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListClustersResponseContent'
          description: ListClusters 200 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Cluster Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.cluster_operations_controller
    post:
      description: Create a managed cluster in a given region.
      operationId: create_cluster
      parameters:
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      - description: "Identifies one or more config validators to suppress. Format:\
          \ (ALL|type:[A-Za-z0-9]+)"
        explode: true
        in: query
        name: suppressValidators
        required: false
        schema:
          description: "Identifies one or more config validators to suppress. Format:\
            \ (ALL|type:[A-Za-z0-9]+)"
          items:
            pattern: "^(ALL|type:[A-Za-z0-9]+)$"
            type: string
          type: array
          uniqueItems: true
        style: form
      - description: Min validation level that will cause the creation to fail. (Defaults
          to 'ERROR'.)
        explode: true
        in: query
        name: validationFailureLevel
        required: false
        schema:
          $ref: '#/components/schemas/ValidationLevel'
        style: form
      - description: Only perform request validation without creating any resource.
          May be used to validate the cluster configuration. (Defaults to 'false'.)
        explode: true
        in: query
        name: dryrun
        required: false
        schema:
          description: Only perform request validation without creating any resource.
            May be used to validate the cluster configuration. (Defaults to 'false'.)
          type: boolean
        style: form
      - description: When set it automatically initiates a cluster stack rollback
          on failures. (Defaults to 'true'.)
        explode: true
        in: query
        name: rollbackOnFailure
        required: false
        schema:
          description: When set it automatically initiates a cluster stack rollback
            on failures. (Defaults to 'true'.)
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClusterRequestContent'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClusterResponseContent'
          description: CreateCluster 202 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClusterBadRequestExceptionResponseContent'
          description: CreateClusterBadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictExceptionResponseContent'
          description: ConflictException 409 response
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryrunOperationExceptionResponseContent'
          description: DryrunOperationException 412 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Cluster Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.cluster_operations_controller
  /v3/clusters/{clusterName}:
    delete:
      description: Initiate the deletion of a cluster.
      operationId: delete_cluster
      parameters:
      - description: Name of the cluster
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          description: Name of the cluster
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          type: string
        style: simple
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteClusterResponseContent'
          description: DeleteCluster 202 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundExceptionResponseContent'
          description: NotFoundException 404 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Cluster Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.cluster_operations_controller
    get:
      description: Get detailed information about an existing cluster.
      operationId: describe_cluster
      parameters:
      - description: Name of the cluster
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          description: Name of the cluster
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          type: string
        style: simple
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribeClusterResponseContent'
          description: DescribeCluster 200 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestExceptionResponseContent'
          description: BadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundExceptionResponseContent'
          description: NotFoundException 404 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Cluster Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.cluster_operations_controller
    put:
      description: Update a cluster managed in a given region.
      operationId: update_cluster
      parameters:
      - description: Name of the cluster
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          description: Name of the cluster
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          type: string
        style: simple
      - description: "Identifies one or more config validators to suppress. Format:\
          \ (ALL|type:[A-Za-z0-9]+)"
        explode: true
        in: query
        name: suppressValidators
        required: false
        schema:
          description: "Identifies one or more config validators to suppress. Format:\
            \ (ALL|type:[A-Za-z0-9]+)"
          items:
            pattern: "^(ALL|type:[A-Za-z0-9]+)$"
            type: string
          type: array
          uniqueItems: true
        style: form
      - description: Min validation level that will cause the update to fail. (Defaults
          to 'ERROR'.)
        explode: true
        in: query
        name: validationFailureLevel
        required: false
        schema:
          $ref: '#/components/schemas/ValidationLevel'
        style: form
      - description: AWS Region that the operation corresponds to.
        explode: true
        in: query
        name: region
        required: false
        schema:
          description: AWS Region that the operation corresponds to.
          type: string
        style: form
      - description: Only perform request validation without creating any resource.
          May be used to validate the cluster configuration and update requirements.
          (Defaults to 'false'.)
        explode: true
        in: query
        name: dryrun
        required: false
        schema:
          description: Only perform request validation without creating any resource.
            May be used to validate the cluster configuration and update requirements.
            (Defaults to 'false'.)
          type: boolean
        style: form
      - description: Force update by ignoring the update validation errors. (Defaults
          to 'false'.)
        explode: true
        in: query
        name: forceUpdate
        required: false
        schema:
          description: Force update by ignoring the update validation errors. (Defaults
            to 'false'.)
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClusterRequestContent'
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateClusterResponseContent'
          description: UpdateCluster 202 response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateClusterBadRequestExceptionResponseContent'
          description: UpdateClusterBadRequestException 400 response
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedClientErrorResponseContent'
          description: UnauthorizedClientError 401 response
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundExceptionResponseContent'
          description: NotFoundException 404 response
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictExceptionResponseContent'
          description: ConflictException 409 response
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryrunOperationExceptionResponseContent'
          description: DryrunOperationException 412 response
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitExceededExceptionResponseContent'
          description: LimitExceededException 429 response
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServiceExceptionResponseContent'
          description: InternalServiceException 500 response
      tags:
      - Cluster Operations
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: "arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ParallelClusterFunction.Arn}/invocations"
        credentials:
          Fn::Sub: "${APIGatewayExecutionRole.Arn}"
        payloadFormatVersion: "2.0"
      x-openapi-router-controller: pcluster.api.controllers.cluster_operations_controller
...
components:
  schemas:
...
    BadRequestExceptionResponseContent:
      description: This exception is thrown when a client calls an API with wrong
        parameters.
      properties:
        message:
          title: message
          type: string
      title: BadRequestExceptionResponseContent
      type: object
...
    ClusterStatus:
      enum:
      - CREATE_IN_PROGRESS
      - CREATE_FAILED
      - CREATE_COMPLETE
      - DELETE_IN_PROGRESS
      - DELETE_FAILED
      - DELETE_COMPLETE
      - UPDATE_IN_PROGRESS
      - UPDATE_COMPLETE
      - UPDATE_FAILED
      title: ClusterStatus
      type: string
    ClusterStatusFilteringOption:
      enum:
      - CREATE_IN_PROGRESS
      - CREATE_FAILED
      - CREATE_COMPLETE
      - DELETE_IN_PROGRESS
      - DELETE_FAILED
      - UPDATE_IN_PROGRESS
      - UPDATE_COMPLETE
      - UPDATE_FAILED
      title: ClusterStatusFilteringOption
      type: string
...
    ConflictExceptionResponseContent:
      description: This exception is thrown when a client request to create/modify
        content would result in a conflict.
      properties:
        message:
          title: message
          type: string
      title: ConflictExceptionResponseContent
      type: object
    CreateClusterBadRequestExceptionResponseContent:
      description: This exception is thrown when a client calls the CreateCluster
        API with an invalid request. This includes an error due to invalid cluster
        configuration.
      properties:
        message:
          title: message
          type: string
        configurationValidationErrors:
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: configurationValidationErrors
          type: array
      title: CreateClusterBadRequestExceptionResponseContent
      type: object
    CreateClusterRequestContent:
      example:
        clusterName: clusterName
        clusterConfiguration: clusterConfiguration
      properties:
        clusterName:
          description: Name of the cluster that will be created.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          title: clusterName
          type: string
        clusterConfiguration:
          description: Cluster configuration as a YAML document.
          title: clusterConfiguration
          type: string
      required:
      - clusterConfiguration
      - clusterName
      title: CreateClusterRequestContent
      type: object
    CreateClusterResponseContent:
      example:
        cluster:
          scheduler:
            metadata:
              name: name
              version: version
            type: type
          clusterName: clusterName
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          region: region
          version: version
          clusterStatus: null
        validationMessages:
        - level: null
          id: id
          type: type
          message: message
        - level: null
          id: id
          type: type
          message: message
      properties:
        cluster:
          $ref: '#/components/schemas/ClusterInfoSummary'
        validationMessages:
          description: List of messages collected during cluster config validation
            whose level is lower than the 'validationFailureLevel' set by the user.
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: validationMessages
          type: array
      required:
      - cluster
      title: CreateClusterResponseContent
      type: object
    DeleteClusterResponseContent:
      example:
        cluster:
          scheduler:
            metadata:
              name: name
              version: version
            type: type
          clusterName: clusterName
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          region: region
          version: version
          clusterStatus: null
      properties:
        cluster:
          $ref: '#/components/schemas/ClusterInfoSummary'
      required:
      - cluster
      title: DeleteClusterResponseContent
      type: object
...
    DescribeClusterResponseContent:
      example:
        failures:
        - failureCode: failureCode
          failureReason: failureReason
        - failureCode: failureCode
          failureReason: failureReason
        creationTime: 2000-01-23T04:56:07.000+00:00
        loginNodes:
          unhealthyNodes: 6
          address: address
          scheme: scheme
          healthyNodes: 0
          status: null
        version: version
        clusterConfiguration:
          url: url
        tags:
        - value: value
          key: key
        - value: value
          key: key
        scheduler:
          metadata:
            name: name
            version: version
          type: type
        cloudFormationStackStatus: null
        clusterName: clusterName
        computeFleetStatus: null
        cloudformationStackArn: cloudformationStackArn
        lastUpdatedTime: 2000-01-23T04:56:07.000+00:00
        region: region
        clusterStatus: null
        headNode:
          launchTime: 2000-01-23T04:56:07.000+00:00
          instanceId: instanceId
          publicIpAddress: publicIpAddress
          instanceType: instanceType
          state: null
          privateIpAddress: privateIpAddress
      properties:
        clusterName:
          description: Name of the cluster.
          pattern: "^[a-zA-Z][a-zA-Z0-9-]+$"
          title: clusterName
          type: string
        region:
          description: AWS region where the cluster is created.
          title: region
          type: string
        version:
          description: ParallelCluster version used to create the cluster.
          title: version
          type: string
        cloudFormationStackStatus:
          $ref: '#/components/schemas/CloudFormationStackStatus'
        clusterStatus:
          $ref: '#/components/schemas/ClusterStatus'
        scheduler:
          $ref: '#/components/schemas/Scheduler'
        cloudformationStackArn:
          description: ARN of the main CloudFormation stack.
          title: cloudformationStackArn
          type: string
        creationTime:
          description: Timestamp representing the cluster creation time.
          format: date-time
          title: creationTime
          type: string
        lastUpdatedTime:
          description: Timestamp representing the last cluster update time.
          format: date-time
          title: lastUpdatedTime
          type: string
        clusterConfiguration:
          $ref: '#/components/schemas/ClusterConfigurationStructure'
        computeFleetStatus:
          $ref: '#/components/schemas/ComputeFleetStatus'
        tags:
          description: Tags associated with the cluster.
          items:
            $ref: '#/components/schemas/Tag'
          title: tags
          type: array
        headNode:
          $ref: '#/components/schemas/EC2Instance'
        loginNodes:
          $ref: '#/components/schemas/LoginNodesPool'
        failures:
          description: Failures array containing failures reason and code when the
            stack is in CREATE_FAILED status.
          items:
            $ref: '#/components/schemas/Failure'
          title: failures
          type: array
      required:
      - cloudFormationStackStatus
      - cloudformationStackArn
      - clusterConfiguration
      - clusterName
      - clusterStatus
      - computeFleetStatus
      - creationTime
      - lastUpdatedTime
      - region
      - tags
      - version
      title: DescribeClusterResponseContent
      type: object
...
    DryrunOperationExceptionResponseContent:
      description: Communicates that the operation would have succeeded without the
        dryrun flag.
      properties:
        message:
          title: message
          type: string
        changeSet:
          description: List of configuration changes requested by the operation.
          items:
            $ref: '#/components/schemas/Change'
          title: changeSet
          type: array
        validationMessages:
          description: List of messages collected during cluster config validation
            whose level is lower than the 'validationFailureLevel' set by the user.
          items:
            $ref: '#/components/schemas/ConfigValidationMessage'
          title: validationMessages
          type: array
      title: DryrunOperationExceptionResponseContent
      type: object
...
    InternalServiceExceptionResponseContent:
      description: This exception is thrown on an unhandled service error.
      properties:
        message:
          title: message
          type: string
      title: InternalServiceExceptionResponseContent
      type: object
    LimitExceededExceptionResponseContent:
      description: The client is sending more than the allowed number of requests
        per unit of time.
      properties:
        message:
          title: message
          type: string
      title: LimitExceededExceptionResponseContent
      type: object
...
    ListClustersResponseContent:
      example:
        nextToken: nextToken
        clusters:
        - scheduler:
            metadata:
              name: name
              version: version
            type: type
          clusterName: clusterName
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          region: region
          version: version
          clusterStatus: null
        - scheduler:
            metadata:
              name: name
              version: version
            type: type
          clusterName: clusterName
          cloudformationStackStatus: null
          cloudformationStackArn: cloudformationStackArn
          region: region
          version: version
          clusterStatus: null
      properties:
        nextToken:
          description: Token to use for paginated requests.
          title: nextToken
          type: string
        clusters:
          items:
            $ref: '#/components/schemas/ClusterInfoSummary'
          title: clusters
          type: array
      required:
      - clusters
      title: ListClustersResponseContent
      type: object
...
    NotFoundExceptionResponseContent:
      description: This exception is thrown when the requested entity is not found.
      properties:
        message:
          title: message
          type: string
      title: NotFoundExceptionResponseContent
      type: object
...
    UnauthorizedClientErrorResponseContent:
      description: This exception is thrown when the client is not authorized to perform
        an action.
      properties:
        message:
          title: message
          type: string
      title: UnauthorizedClientErrorResponseContent
      type: object
...
    ValidationLevel:
      enum:
      - INFO
      - WARNING
      - ERROR
      title: ValidationLevel
      type: string
  securitySchemes:
    aws.auth.sigv4:
      description: AWS Signature Version 4 authentication
      in: header
      name: Authorization
      type: apiKey
      x-amazon-apigateway-authtype: awsSigv4
      x-apikeyInfoFunc: pcluster.api.controllers.security_controller_.info_from_aws.auth.sigv4